move common fixtures to conftest
